requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
altair==6.0.0
    # via streamlit
anyio==4.12.1
    # via
    #   httpx
    #   jupyter-server
argon2-cffi==25.1.0
    # via jupyter-server
argon2-cffi-bindings==25.1.0
    # via argon2-cffi
arrow==1.4.0
    # via isoduration
asttokens==3.0.1
    # via stack-data
async-lru==2.1.0
    # via jupyterlab
attrs==25.4.0
    # via
    #   jsonschema
    #   referencing
babel==2.18.0
    # via jupyterlab-server
beautifulsoup4==4.14.3
    # via nbconvert
bleach[css]==6.3.0
    # via nbconvert
blinker==1.9.0
    # via streamlit
cachetools==6.2.6
    # via streamlit
certifi==2026.1.4
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via argon2-cffi-bindings
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via streamlit
comm==0.2.3
    # via
    #   ipykernel
    #   ipywidgets
debugpy==1.8.20
    # via ipykernel
decorator==5.2.1
    # via ipython
defusedxml==0.7.1
    # via nbconvert
executing==2.2.1
    # via stack-data
fastjsonschema==2.21.2
    # via nbformat
fqdn==1.5.1
    # via jsonschema
gitdb==4.0.12
    # via gitpython
gitpython==3.1.46
    # via streamlit
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via jupyterlab
idna==3.11
    # via
    #   anyio
    #   httpx
    #   jsonschema
    #   requests
ipykernel==7.1.0
    # via
    #   jupyter
    #   jupyter-console
    #   jupyterlab
ipython==9.10.0
    # via
    #   ipykernel
    #   ipywidgets
    #   jupyter-console
ipython-pygments-lexers==1.1.1
    # via ipython
ipywidgets==8.1.8
    # via jupyter
isoduration==20.11.0
    # via jsonschema
jedi==0.19.2
    # via ipython
jinja2==3.1.6
    # via
    #   altair
    #   jupyter-server
    #   jupyterlab
    #   jupyterlab-server
    #   nbconvert
    #   pydeck
joblib==1.5.3
    # via scikit-learn
json5==0.13.0
    # via jupyterlab-server
jsonpointer==3.0.0
    # via jsonschema
jsonschema[format-nongpl]==4.26.0
    # via
    #   altair
    #   jupyter-events
    #   jupyterlab-server
    #   nbformat
jsonschema-specifications==2025.9.1
    # via jsonschema
jupyter==1.1.1
    # via -r requirements.in
jupyter-client==8.8.0
    # via
    #   ipykernel
    #   jupyter-console
    #   jupyter-server
    #   nbclient
jupyter-console==6.6.3
    # via jupyter
jupyter-core==5.9.1
    # via
    #   ipykernel
    #   jupyter-client
    #   jupyter-console
    #   jupyter-server
    #   jupyterlab
    #   nbclient
    #   nbconvert
    #   nbformat
jupyter-events==0.12.0
    # via jupyter-server
jupyter-lsp==2.3.0
    # via jupyterlab
jupyter-server==2.17.0
    # via
    #   jupyter-lsp
    #   jupyterlab
    #   jupyterlab-server
    #   notebook
    #   notebook-shim
jupyter-server-terminals==0.5.4
    # via jupyter-server
jupyterlab==4.5.3
    # via
    #   jupyter
    #   notebook
jupyterlab-pygments==0.3.0
    # via nbconvert
jupyterlab-server==2.28.0
    # via
    #   jupyterlab
    #   notebook
jupyterlab-widgets==3.0.16
    # via ipywidgets
lark==1.3.1
    # via rfc3987-syntax
markupsafe==3.0.3
    # via
    #   jinja2
    #   nbconvert
matplotlib-inline==0.2.1
    # via
    #   ipykernel
    #   ipython
mistune==3.2.0
    # via nbconvert
narwhals==2.16.0
    # via altair
nbclient==0.10.4
    # via nbconvert
nbconvert==7.17.0
    # via
    #   jupyter
    #   jupyter-server
nbformat==5.10.4
    # via
    #   jupyter-server
    #   nbclient
    #   nbconvert
nest-asyncio==1.6.0
    # via ipykernel
notebook==7.5.3
    # via jupyter
notebook-shim==0.2.4
    # via
    #   jupyterlab
    #   notebook
numpy==2.4.2
    # via
    #   -r requirements.in
    #   pandas
    #   pydeck
    #   scikit-learn
    #   scipy
    #   streamlit
packaging==26.0
    # via
    #   altair
    #   ipykernel
    #   jupyter-events
    #   jupyter-server
    #   jupyterlab
    #   jupyterlab-server
    #   nbconvert
    #   streamlit
pandas==2.3.3
    # via
    #   -r requirements.in
    #   streamlit
pandocfilters==1.5.1
    # via nbconvert
parso==0.8.5
    # via jedi
pexpect==4.9.0
    # via ipython
pillow==12.1.0
    # via streamlit
platformdirs==4.5.1
    # via jupyter-core
prometheus-client==0.24.1
    # via jupyter-server
prompt-toolkit==3.0.52
    # via
    #   ipython
    #   jupyter-console
protobuf==6.33.5
    # via streamlit
psutil==7.2.2
    # via ipykernel
ptyprocess==0.7.0
    # via
    #   pexpect
    #   terminado
pure-eval==0.2.3
    # via stack-data
pyarrow==23.0.0
    # via streamlit
pycparser==3.0
    # via cffi
pydeck==0.9.1
    # via streamlit
pygments==2.19.2
    # via
    #   ipython
    #   ipython-pygments-lexers
    #   jupyter-console
    #   nbconvert
python-dateutil==2.9.0.post0
    # via
    #   arrow
    #   jupyter-client
    #   pandas
python-json-logger==4.0.0
    # via jupyter-events
pytz==2025.2
    # via pandas
pyyaml==6.0.3
    # via jupyter-events
pyzmq==27.1.0
    # via
    #   ipykernel
    #   jupyter-client
    #   jupyter-console
    #   jupyter-server
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
    #   jupyter-events
requests==2.32.5
    # via
    #   jupyterlab-server
    #   streamlit
rfc3339-validator==0.1.4
    # via
    #   jsonschema
    #   jupyter-events
rfc3986-validator==0.1.1
    # via
    #   jsonschema
    #   jupyter-events
rfc3987-syntax==1.1.0
    # via jsonschema
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
scikit-learn==1.8.0
    # via -r requirements.in
scipy==1.17.0
    # via
    #   -r requirements.in
    #   scikit-learn
send2trash==2.1.0
    # via jupyter-server
six==1.17.0
    # via
    #   python-dateutil
    #   rfc3339-validator
smmap==5.0.2
    # via gitdb
soupsieve==2.8.3
    # via beautifulsoup4
stack-data==0.6.3
    # via ipython
streamlit==1.54.0
    # via -r requirements.in
tenacity==9.1.3
    # via streamlit
terminado==0.18.1
    # via
    #   jupyter-server
    #   jupyter-server-terminals
threadpoolctl==3.6.0
    # via scikit-learn
tinycss2==1.4.0
    # via bleach
toml==0.10.2
    # via streamlit
tornado==6.5.4
    # via
    #   ipykernel
    #   jupyter-client
    #   jupyter-server
    #   jupyterlab
    #   notebook
    #   streamlit
    #   terminado
traitlets==5.14.3
    # via
    #   ipykernel
    #   ipython
    #   ipywidgets
    #   jupyter-client
    #   jupyter-console
    #   jupyter-core
    #   jupyter-events
    #   jupyter-server
    #   jupyterlab
    #   matplotlib-inline
    #   nbclient
    #   nbconvert
    #   nbformat
typing-extensions==4.15.0
    # via
    #   altair
    #   anyio
    #   beautifulsoup4
    #   referencing
    #   streamlit
tzdata==2025.3
    # via
    #   arrow
    #   pandas
uri-template==1.3.0
    # via jsonschema
urllib3==2.6.3
    # via requests
watchdog==6.0.0
    # via streamlit
wcwidth==0.5.3
    # via prompt-toolkit
webcolors==25.10.0
    # via jsonschema
webencodings==0.5.1
    # via
    #   bleach
    #   tinycss2
websocket-client==1.9.0
    # via jupyter-server
widgetsnbextension==4.0.15
    # via ipywidgets

# The following packages are considered to be unsafe in a requirements file:
# setuptools
